Update start_tracer function to accept url
The start_tracer function in the lib.rs file has been updated to accept a URL as a parameter. This makes the function more modular, as it can now connect to any specified URL, rather than being hard-coded to connect to a specific address. The version in the Cargo.toml file was also incremented to reflect this change.