Add retry logic to log export and update package version
The updated code introduces a loop in the log export process to handle potential errors. In case of a failure when attempting to send the logs, an error message is printed, the thread sleeps for 1 second before trying again. The version of the package in Cargo.toml is also updated from