Refactor log sending code and update package version
Place log sending code within a new external loop to retry on failure. The existing loop logic is adjusted so that every time a new log is generated, it is attempted to send immediately rather than cloned. Errors during the sending process are now logged to the standard error instead of standard output. Also, the version of the 'telescope-client' package in Cargo.toml is incremented from "0.1.11" to "0.1.12".